idem

+++ src/unlam/paradigmas/tp2/hechizos/HechizoAtaque.java
@@ -22,7 +22,7 @@ public class HechizoAtaque implements Hechizo{
 	}
 	*/
 	
-	public int ejecutar(Personaje objetivo, int manaActualPersonaje) {
+	public double ejecutar(Personaje objetivo, double manaActualPersonaje) {
 		return 0;
 		
 	}
